requirements.txt
python-dotenv~=0.21.1
py-cord